--- src/_legacy_components/atoms/ButtonArrow.tsx
@@ -1,33 +1,0 @@
-import { faArrowLeft, faArrowRight } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { cls } from '../../utils/common.utils';
-import { NEXT, PREV } from '../../constants/constants';
-
-interface BtnArrowProps {
-  direction: typeof PREV | typeof NEXT;
-  onClick: () => void;
-  className?: string;
-}
-
-export default function BtnArrow({
-  direction,
-  onClick,
-  className,
-}: BtnArrowProps) {
-  return (
-    <button
-      className={cls(
-        'btn-menu rounded-md border bg-white p-1',
-        className ?? ''
-      )}
-      onClick={onClick}
-      type="button"
-    >
-      {direction === PREV ? (
-        <FontAwesomeIcon icon={faArrowLeft} fontSize={14} />
-      ) : (
-        <FontAwesomeIcon icon={faArrowRight} fontSize={14} />
-      )}
-    </button>
-  );
-}
