+++ src/pages/timetable/components/SettingOfTimetable/index.ts
@@ -1,0 +1,3 @@
+import SettingOfTimetable from './SettingOfTimetable';
+
+export { SettingOfTimetable };
