+++ src/types/store.types.ts
@@ -1,0 +1,5 @@
+export type HiddenUsersArr = number[];
+
+export type HiddenUsersSet = Set<number>;
+
+export type ClinicIdAndHiddenUsers = [number, number[]];
